# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Deketna business application",
        "title": "Deketna API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of all orders with buyer details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Orders"
                ],
                "summary": "Admin View Orders",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by order status (e.g., pending, completed, cancelled)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of all orders with details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.PaginatedAdminOrdersResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/order": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new order from the buyer's cart, validate stock, deduct quantities, and clear the cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "Place Order",
                "responses": {
                    "200": {
                        "description": "Order placed successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "order_id": {
                                                    "type": "integer"
                                                },
                                                "total_amount": {
                                                    "type": "number"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request: Cart is empty or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of orders placed by the authenticated buyer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Orders"
                ],
                "summary": "View Orders",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of orders with details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.PaginatedOrdersResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.AdminOrderResponse": {
            "type": "object",
            "properties": {
                "buyer": {
                    "$ref": "#/definitions/admin.OrderBuyerResponse"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.OrderItemResponse"
                    }
                },
                "order_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "total_amount": {
                    "type": "number",
                    "example": 75.5
                }
            }
        },
...
        "admin.OrderBuyerResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "buyer@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "phone": {
                    "type": "string",
                    "example": "123456789"
                }
            }
        },
        "admin.OrderItemResponse": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number",
                    "example": 25
                },
                "product_name": {
                    "type": "string",
                    "example": "Product A"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "admin.PaginatedAdminOrdersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.AdminOrderResponse"
                    }
                },
                "isNext": {
                    "type": "boolean",
                    "example": true
                },
                "isPrev": {
                    "type": "boolean",
                    "example": false
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "totalItem": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
...
        "helper.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "helper.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/helper.ErrorDetail"
                        }
                    ]
                }
            }
        },
        "helper.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)"
                },
                "message": {
                    "description": "Description of the operation",
                    "type": "string"
                }
            }
        },
...
        "user.OrderItemResponse": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number",
                    "example": 25
                },
                "product_name": {
                    "type": "string",
                    "example": "Product A"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "user.OrderResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.OrderItemResponse"
                    }
                },
                "order_id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "total_amount": {
                    "type": "number",
                    "example": 75.5
                }
            }
        },
        "user.PaginatedOrdersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "List of orders",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.OrderResponse"
                    }
                },
                "isNext": {
                    "description": "Whether there is a next page",
                    "type": "boolean",
                    "example": true
                },
                "isPrev": {
                    "description": "Whether there is a previous page",
                    "type": "boolean",
                    "example": false
                },
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "description": "Current page number",
                    "type": "integer",
                    "example": 1
                },
                "totalItem": {
                    "description": "Total number of items",
                    "type": "integer",
                    "example": 25
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer \u003ctoken\u003e\" (e.g., \"Bearer abc123\") as the value.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}